requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-genai>=0.3.0
pdf2image==1.17.0
python-docx==1.1.0
pillow==10.2.0
requests==2.31.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
